Make it so the script can be re-run and not overwrite existing files

+++ setup.sh
@@ -1,4 +1,19 @@
 #!/usr/bin/env bash
-ln -s `pwd`/.config ~/.config
-ln -s `pwd`/.tmux.conf ~/.tmux.conf
-ln -s `pwd`/.vimrc ~/.vimrc
+echo "==> Linking Files"
+if [ ! -d ~/.config ]; then
+	ln -s `pwd`/.config ~/.config
+else
+	echo ".config exists, skipping"
+fi
+
+if [ ! -f ~/.tmux.conf ]; then
+	ln -s `pwd`/.tmux.conf ~/.tmux.conf
+else
+	echo ".tmux.conf exists, skipping"
+fi
+
+if [ ! -f ~/.vimrc ]; then
+	ln -s `pwd`/.vimrc ~/.vimrc
+else
+	echo ".vimrc exists, skipping"
+fi
